run rema without 2023 biomass
per discussion with J Sullivan, wanted to investigate impact of the really odd 2023 EGOA measurement (since the apportionment model doesn't fit that upswing). As she suggested, when we run the REMA model through 2021 (all areas) the process error in EGOA is reduced (0.161 vs 0.177), meaning that the model *is* responding to the outrageousness of that value -- but not to the degree of fitting it. As Jane said, "I don't think you'd be able to simulate that observation!". This is precisely the situation that REMA was designed to accomodate.