edit figure to show 2021 valeus
the MCMC analysis done in 2021 did not save the posteriors for age-2 recruits nor F (I had to manually edit the .tpl to get these to show up for the 2023 analysis). So instead we are reading those two params from the bio_rec_F.csv, and the bottom two rows have uncertainty from mceval.csv; the caption has been updated to clarify.