check 2021 rec dev posteriors
I got confused about how the terminal rec-devs behave; I had it in my head they were forced to zero (since in 2023 the last two years are that way), but this plot shows that in 2021 that wasn't true for the terminal two years. Looking at the code, it's simply that rec-devs are penalized throughout the time series, which explains the mean-tendency with add'l years.